+++ packages/realm/src/index.ts
@@ -139,6 +139,7 @@ export {
   SubscriptionOptions,
   SubscriptionSet,
   SubscriptionsState,
+  Sync,
   SyncConfiguration,
   SyncError,
   SyncSession,