+++ templates/expo-template-js/app/models/index.js
@@ -1,6 +1,0 @@
-import { createRealmContext } from "@realm/react";
-import { Task } from "./Task";
-
-export const TaskRealmContext = createRealmContext({
-  schema: [Task],
-});