+++ templates/react-native-template-realm-js/template.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  placeholderName: 'RealmJsTemplate',
-  titlePlaceholder: 'RealmJsTemplate',
-  templateDir: './template',
-};